Correccion activdad 8

+++ TP 1/Actividad 8.C
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int monto_en_pesos;
-int precio_dolar = 15.40;
+float monto_en_pesos;
+float precio_dolar = 15.40;
 
 main()
 {
     printf("Ingrese el monto en pesos: \n");
-    scanf("%i", &monto_en_pesos);
-    printf("El monto ingresado equivale a %d dolares", monto_en_pesos/precio_dolar);
+    scanf("%f", &monto_en_pesos);
+    printf("El monto ingresado equivale a %f dolares", monto_en_pesos / precio_dolar);
     return 0;
 }
